migrate2GITEA/requirements.txt
#pip install -r requirements.txt
gitpython==3.1.42
httpx==0.27.0
requests==2.25.1
